Phase 8: finalize UI wiring + zero-warning build
Phase 8 (UI) complete and user-approved via UIRenderProbe renders. The
editor/asset groundwork built ahead of schedule needed no DSP-wiring
changes post-Phase-7 — all 15 APVTS params have live editor attachments
feeding the per-revision DSP graphs.

UI tweak:
- ThreePositionSwitch revision labels (V1 EARLY/V1 LATE/V2) stretched 30%
  wider via a horizontal-only AffineTransform anchored at the label's
  left edge (track/lever position unchanged); PluginEditor kSwitchW ratio
  widened (2.7x->3.4x) so the stretched text doesn't clip.

Warning cleanup (build now warning-clean from scratch, all -Wall -Wextra
+ JUCE recommended flags):
- -Wfloat-equal x14: new src/utils/ChangeGate.h nalr::changed() does the
  pot-didn't-move dirty-flag compare via memcmp (exact-equality is the
  correct semantic; bit-identical to !=) across V1Early/V1Late/V2 DSP.
- -Wsign-conversion x7: V1EarlyDriveClipRecovery kNumOs/osIndex() now
  size_t, loop counters size_t — native unsigned array indices, no casts.
- -Wunused-parameter: unnamed the unused `highlighted` in
  PedalLookAndFeel::drawButtonBackground.
- -Wswitch-enum: PedalAssets::texture() lists Revision::v2 explicitly.
- -Wdeprecated-declarations: offline_render migrated to JUCE 8
  createWriterFor(unique_ptr<OutputStream>&, AudioFormatWriterOptions).

All behavior-preserving; ctest 19/19 green.

+++ src/dsp/V1EarlyDriveClipRecovery.h
@@ -39,9 +39,9 @@ public:
     void prepare(double baseFs, int maxBlock)
     {
         baseSampleRate = baseFs;
-        for (int i = 0; i < kNumOs; ++i)
+        for (size_t i = 0; i < kNumOs; ++i)
         {
-            const size_t stages = (size_t) (i + 1); // os[0]=2x(1 stage), os[1]=4x, os[2]=8x
+            const size_t stages = i + 1; // os[0]=2x(1 stage), os[1]=4x, os[2]=8x
             os[i] = std::make_unique<juce::dsp::Oversampling<double>>(
                 1, stages, juce::dsp::Oversampling<double>::filterHalfBandFIREquiripple,
                 /*isMaxQuality*/ true, /*useIntegerLatency*/ true);
@@ -63,7 +63,7 @@ public:
         drive.reset();
         railClip.reset();
         recovery.reset();
-        for (int i = 0; i < kNumOs; ++i)
+        for (size_t i = 0; i < kNumOs; ++i)
             if (os[i] != nullptr)
                 os[i]->reset();
     }
@@ -110,9 +110,9 @@ public:
     int getActiveFactor() const noexcept { return activeFactor; }
 
 private:
-    static constexpr int kNumOs = 3; // 2x, 4x, 8x
+    static constexpr size_t kNumOs = 3; // 2x, 4x, 8x
 
-    static int osIndex(int factor) noexcept { return factor == 2 ? 0 : (factor == 4 ? 1 : 2); }
+    static size_t osIndex(int factor) noexcept { return factor == 2 ? 0u : (factor == 4 ? 1u : 2u); }
 
     void applyFactor(int factor) noexcept
     {
